requirements.txt
streamlit
pandas
numpy
tensorflow
pillow
opencv-python-headless
scikit-learn
pymysql
plotly
dotenv